legacy/Agents/IT22106056_Dhanaga_Agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
PyMuPDF==1.23.8
python-docx==1.1.0
requests==2.31.0
spacy==3.7.2
word2number==1.1
python-dotenv==1.0.0
